+++ examples/SampleApp/src/components/UnreadCountBadge.tsx
@@ -16,7 +16,7 @@ export const ThreadsUnreadCountBadge: React.FC = () => {
     return null;
   }
 
-  return <BadgeNotification count={unreadCount} size='md' type='primary' />;
+  return <BadgeNotification count={unreadCount} size='sm' type='primary' />;
 };
 
 export const ChannelsUnreadCountBadge: React.FC = () => {
@@ -55,5 +55,5 @@ export const ChannelsUnreadCountBadge: React.FC = () => {
     return null;
   }
 
-  return <BadgeNotification count={unreadCount} size='md' type='primary' />;
+  return <BadgeNotification count={unreadCount} size='sm' type='primary' />;
 };