+++ examples/SampleApp/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## [4.14.7](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v4.14.6...sampleapp@v4.14.7) (2026-08-14)
+
+### Performance Improvements
+
+* introduce offline sync event limit ([#3773](https://github.com/GetStream/stream-chat-react-native/issues/3773)) ([3e56a09](https://github.com/GetStream/stream-chat-react-native/commit/3e56a09bd7acef8b4acbd0c1b69dd4f001697da5))
+
+### Reverts
+
+* Revert "chore: bump stream-chat-js" ([251eb79](https://github.com/GetStream/stream-chat-react-native/commit/251eb79067968f10d9b2f7d46f1dbf0be7b06287))
+* Revert "perf: sync event count" ([eb1022e](https://github.com/GetStream/stream-chat-react-native/commit/eb1022e2550f53dd30ab0185daf53504deaa896f))
+
+## [4.14.6](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v4.14.5...sampleapp@v4.14.6) (2026-08-10)
+
+### Bug Fixes
+
+* recover db gracefully on errors ([#3764](https://github.com/GetStream/stream-chat-react-native/issues/3764)) ([e8c917c](https://github.com/GetStream/stream-chat-react-native/commit/e8c917c05a0567a9b0d3f60e5e7c5e9763d948b4))
+
+## [4.14.5](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v4.14.4...sampleapp@v4.14.5) (2026-08-06)
+
+### Bug Fixes
+
+* bump nitro-sound to latest on SampleApp ([#3757](https://github.com/GetStream/stream-chat-react-native/issues/3757)) ([f9848f5](https://github.com/GetStream/stream-chat-react-native/commit/f9848f53300abbe3007a37e2bb3f336afed34811))
+* expo clipboard and extend with cb api ([#3758](https://github.com/GetStream/stream-chat-react-native/issues/3758)) ([2534581](https://github.com/GetStream/stream-chat-react-native/commit/253458114aa1bd23b17ad3a920813ce6149e6388))
+
 ## [4.14.4](https://github.com/GetStream/stream-chat-react-native/compare/sampleapp@v4.14.3...sampleapp@v4.14.4) (2026-07-31)
 
 ### Bug Fixes